Split ICA decoding modules
