--- src/features/messages/ica.ts
@@ -1,1398 +1,0 @@
-import { chainAddresses } from '@hyperlane-xyz/registry';
-import type { ChainMap } from '@hyperlane-xyz/sdk';
-import {
-  ProtocolType,
-  addressToBytes32,
-  bytes32ToAddress,
-  fromWei,
-  shortenAddress,
-  isEVMLike,
-  strip0x,
-} from '@hyperlane-xyz/utils';
-import { useQuery } from '@tanstack/react-query';
-import { BigNumber, Contract, providers, utils } from 'ethers';
-import { useMemo } from 'react';
-import { useQuery as useUrqlQuery } from 'urql';
-
-import { useChainMetadataResolver } from '../../metadataStore';
-import { useReadyMultiProvider, useRegistry } from '../../store';
-import { IcaCall } from '../../types';
-import { formatAmountWithCommas } from '../../utils/amount';
-import { logger } from '../../utils/logger';
-import { useScrapedDomains } from '../chains/queries/useScrapedChains';
-import { isPiChain } from '../chains/utils';
-import type { ExplorerMultiProvider } from '../hyperlane/sdkRuntime';
-import { PiQueryType, fetchMessagesFromPiChain } from './pi-queries/fetchPiChainMessages';
-import { MessageIdentifierType, buildMessageQuery } from './queries/build';
-import { MessagesStubQueryResult } from './queries/fragments';
-import { parseMessageStubResult } from './queries/parse';
-
-/**
- * ICA Message Types (from InterchainAccountMessage.sol)
- */
-export enum IcaMessageType {
-  CALLS = 0,
-  COMMITMENT = 1,
-  REVEAL = 2,
-}
-
-/**
- * Decoded ICA message with all fields
- */
-export interface DecodedIcaMessage {
-  messageType: IcaMessageType;
-  owner: string; // bytes32 -> address
-  ism: string; // bytes32 -> address
-  salt: string; // bytes32 hex
-  calls: IcaCall[]; // Only present for CALLS type
-  commitment?: string; // Only present for COMMITMENT type
-}
-
-export interface DecodedIcaCallData {
-  functionName: string;
-  summary: string;
-  details?: Array<{ label: string; value: string }>;
-  swap?: {
-    tokenIn: string;
-    tokenOut: string;
-    tokenOutType?: 'token' | 'native';
-    outputAmount?: string;
-    outputAmountKind?: 'exact' | 'minimum';
-    wrappedNativeToken?: string;
-    outputRecipients?: string[];
-  };
-}
-
-// Build a map of chainName -> ICA router address from the registry
-function buildIcaRouterAddressMap(): ChainMap<Address> {
-  const map: ChainMap<Address> = {};
-
-  for (const [chainName, addresses] of Object.entries(chainAddresses)) {
-    const icaRouter = (addresses as Record<string, string>).interchainAccountRouter;
-    if (icaRouter) {
-      map[chainName] = icaRouter;
-    }
-  }
-
-  return map;
-}
-
-// Cached ICA router address map built at module load time
-const ICA_ROUTER_MAP = buildIcaRouterAddressMap();
-
-// Normalize to bytes32 hex so registry entries (EVM-style 0x hex) and
-// resolved message addresses (e.g. base58 for Tron) compare equal when they
-// represent the same on-chain contract.
-function toCanonicalBytes32(addr: Address): string | null {
-  if (!addr) return null;
-  try {
-    return addressToBytes32(addr).toLowerCase();
-  } catch {
-    return null;
-  }
-}
-
-// Get all known ICA router addresses, canonicalized to bytes32 hex
-function getIcaRouterBytes32Set(): Set<string> {
-  const set = new Set<string>();
-  for (const addr of Object.values(ICA_ROUTER_MAP)) {
-    const canonical = toCanonicalBytes32(addr);
-    if (canonical) set.add(canonical);
-  }
-  return set;
-}
-
-/**
- * Check if an address is a known ICA router
- */
-export function isAddressIcaRouter(addr: Address): boolean {
-  if (!addr) return false;
-  try {
-    const canonical = toCanonicalBytes32(addr);
-    if (!canonical) return false;
-    return getIcaRouterBytes32Set().has(canonical);
-  } catch (error) {
-    logger.warn('Error checking if address is ICA router', error, addr);
-    return false;
-  }
-}
-
-/**
- * Check if a message is an ICA message by verifying both sender and recipient
- * are known ICA router addresses
- */
-export function isIcaMessage({
-  sender,
-  recipient,
-}: {
-  sender: Address;
-  recipient: Address;
-}): boolean {
-  const isSenderIca = isAddressIcaRouter(sender);
-  const isRecipIca = isAddressIcaRouter(recipient);
-
-  if (isSenderIca && isRecipIca) return true;
-
-  if (isSenderIca && !isRecipIca) {
-    logger.warn('Msg sender is ICA router but not recipient', sender, recipient);
-  }
-  if (!isSenderIca && isRecipIca) {
-    logger.warn('Msg recipient is ICA router but not sender', recipient, sender);
-  }
-
-  return false;
-}
-
-/**
- * React hook to check if a message is an ICA message
- */
-export function useIsIcaMessage({
-  sender,
-  recipient,
-}: {
-  sender: Address;
-  recipient: Address;
-}): boolean {
-  return useMemo(() => isIcaMessage({ sender, recipient }), [sender, recipient]);
-}
-
-/**
- * Decode an ICA message body.
- *
- * Message formats (from InterchainAccountMessage.sol):
- *
- * CALLS message:
- * - Byte 0: MessageType.CALLS (0x00)
- * - Bytes 1-33: ICA owner (bytes32)
- * - Bytes 33-65: ICA ISM (bytes32)
- * - Bytes 65-97: User Salt (bytes32)
- * - Bytes 97+: ABI-encoded Call[] where Call = (bytes32 to, uint256 value, bytes data)
- *
- * COMMITMENT message:
- * - Byte 0: MessageType.COMMITMENT (0x01)
- * - Bytes 1-33: ICA owner (bytes32)
- * - Bytes 33-65: ICA ISM (bytes32)
- * - Bytes 65-97: User Salt (bytes32)
- * - Bytes 97-129: Commitment (bytes32)
- *
- * REVEAL message:
- * - Byte 0: MessageType.REVEAL (0x02)
- * - Bytes 1-33: ICA ISM (bytes32)
- * - Bytes 33-65: Commitment (bytes32)
- */
-export function decodeIcaBody(body: string): DecodedIcaMessage | null {
-  if (!body) return null;
-
-  try {
-    const bodyHex = strip0x(body);
-
-    // Safe zero check - handles any length payload without throwing
-    if (!bodyHex || /^0*$/.test(bodyHex)) return null;
-
-    // Minimum length to read message type: 1 byte = 2 hex chars
-    if (bodyHex.length < 2) {
-      logger.warn('ICA body too short to read message type');
-      return null;
-    }
-
-    // Parse message type (first byte)
-    const messageType = parseInt(bodyHex.slice(0, 2), 16) as IcaMessageType;
-
-    if (messageType === IcaMessageType.REVEAL) {
-      // REVEAL format: type (1) + ism (32) + commitment (32) = 65 bytes = 130 hex chars
-      if (bodyHex.length < 130) {
-        logger.warn('REVEAL message body too short');
-        return null;
-      }
-
-      const revealIsm = bytes32ToAddress('0x' + bodyHex.slice(2, 66));
-      const revealCommitment = '0x' + bodyHex.slice(66, 130);
-
-      return {
-        messageType,
-        owner: '', // Not present in REVEAL
-        ism: revealIsm,
-        salt: '', // Not present in REVEAL
-        calls: [],
-        commitment: revealCommitment,
-      };
-    }
-
-    // CALLS and COMMITMENT messages have the same prefix format
-    // Minimum length: 1 byte type + 32 bytes owner + 32 bytes ism + 32 bytes salt = 97 bytes = 194 hex chars
-    if (bodyHex.length < 194) {
-      logger.warn('ICA CALLS/COMMITMENT body too short');
-      return null;
-    }
-
-    // Parse owner (bytes 1-33)
-    const ownerBytes32 = '0x' + bodyHex.slice(2, 66);
-    const owner = bytes32ToAddress(ownerBytes32);
-
-    // Parse ISM (bytes 33-65)
-    const ismBytes32 = '0x' + bodyHex.slice(66, 130);
-    const ism = bytes32ToAddress(ismBytes32);
-
-    // Parse salt (bytes 65-97)
-    const salt = '0x' + bodyHex.slice(130, 194);
-
-    if (messageType === IcaMessageType.CALLS) {
-      // Decode the ABI-encoded calls array (bytes 97+)
-      const encodedCalls = '0x' + bodyHex.slice(194);
-
-      // Format: (bytes32 to, uint256 value, bytes data)[]
-      const decoded = utils.defaultAbiCoder.decode(
-        ['tuple(bytes32 to, uint256 value, bytes data)[]'],
-        encodedCalls,
-      );
-
-      const rawCalls = decoded[0] as Array<{
-        to: string;
-        value: BigNumber;
-        data: string;
-      }>;
-
-      const calls: IcaCall[] = rawCalls.map((call) => ({
-        to: bytes32ToAddress(call.to),
-        value: call.value.toString(),
-        data: call.data,
-      }));
-
-      return { messageType, owner, ism, salt, calls };
-    } else if (messageType === IcaMessageType.COMMITMENT) {
-      // Commitment is bytes 97-129
-      if (bodyHex.length < 258) {
-        logger.warn('COMMITMENT message body too short for commitment hash');
-        return null;
-      }
-      const commitment = '0x' + bodyHex.slice(194, 258);
-      return { messageType, owner, ism, salt, calls: [], commitment };
-    }
-
-    // Unknown message type
-    logger.warn('Unknown ICA message type:', messageType);
-    return null;
-  } catch (error) {
-    logger.error('Error decoding ICA body', error);
-    return null;
-  }
-}
-
-/**
- * Get the ICA router address for a given chain
- */
-export function getIcaRouterAddress(chainName: string): Address | undefined {
-  return ICA_ROUTER_MAP[chainName];
-}
-
-/**
- * Compute the ICA address for a given owner on the destination chain.
- * This is a non-hook version for use in non-React contexts like the debugger.
- *
- * @param originDomainId - The origin chain's domain ID
- * @param owner - The owner address on the origin chain
- * @param originRouter - The ICA router address on the origin chain
- * @param destRouter - The ICA router address on the destination chain
- * @param ism - Optional ISM address (uses default if not specified)
- * @param salt - Optional salt (uses zero salt if not specified)
- * @param destProvider - Provider for the destination chain
- * @returns The derived ICA address, or null if computation fails
- */
-export async function computeIcaAddress(
-  originDomainId: number,
-  owner: string,
-  originRouter: string,
-  destRouter: string,
-  ism: string | undefined,
-  salt: string | undefined,
-  destProvider: providers.Provider,
-): Promise<string | null> {
-  try {
-    // eslint-disable-next-line camelcase
-    const { InterchainAccountRouter__factory } = await import('@hyperlane-xyz/core');
-    const router = InterchainAccountRouter__factory.connect(destRouter, destProvider);
-
-    // Use zero address for ISM if not specified (will use default ISM)
-    const ismAddress = ism || '0x0000000000000000000000000000000000000000';
-    const userSalt = salt || '0x' + '0'.repeat(64);
-
-    // Get the ICA address using the contract with salt
-    // Signature: getLocalInterchainAccount(uint32,bytes32,bytes32,address,bytes32)
-    const icaAddress = await router[
-      'getLocalInterchainAccount(uint32,bytes32,bytes32,address,bytes32)'
-    ](
-      originDomainId,
-      addressToBytes32(owner),
-      addressToBytes32(originRouter),
-      ismAddress,
-      userSalt,
-    );
-
-    return icaAddress;
-  } catch (error) {
-    logger.error('Error computing ICA address', error);
-    return null;
-  }
-}
-
-/**
- * Decode the ISM metadata for a REVEAL message to extract the calls.
- *
- * Metadata format (from CommitmentReadIsm.verify):
- * - Bytes 0-20: ICA address
- * - Bytes 20-52: Salt (bytes32)
- * - Bytes 52+: ABI-encoded CallLib.Call[]
- */
-function decodeRevealMetadata(metadata: string): {
-  icaAddress: string;
-  salt: string;
-  calls: IcaCall[];
-} | null {
-  try {
-    const metaHex = strip0x(metadata);
-
-    // Minimum: 20 bytes address + 32 bytes salt = 52 bytes = 104 hex chars
-    if (metaHex.length < 104) {
-      return null;
-    }
-
-    // ICA address (bytes 0-20)
-    const icaAddress = '0x' + metaHex.slice(0, 40);
-
-    // Salt (bytes 20-52)
-    const salt = '0x' + metaHex.slice(40, 104);
-
-    // Calls (bytes 52+)
-    const encodedCalls = '0x' + metaHex.slice(104);
-
-    const decoded = utils.defaultAbiCoder.decode(
-      ['tuple(bytes32 to, uint256 value, bytes data)[]'],
-      encodedCalls,
-    );
-
-    const rawCalls = decoded[0] as Array<{
-      to: string;
-      value: BigNumber;
-      data: string;
-    }>;
-
-    const calls: IcaCall[] = rawCalls.map((call) => ({
-      to: bytes32ToAddress(call.to),
-      value: call.value.toString(),
-      data: call.data,
-    }));
-
-    return { icaAddress, salt, calls };
-  } catch (error) {
-    logger.error('Error decoding reveal metadata', error);
-    return null;
-  }
-}
-
-// Multicall3 canonical address (deployed on 70+ chains at the same address)
-const MULTICALL3_ADDRESS = '0xcA11bde05977b3631167028862bE2a173976CA11';
-
-// Common Multicall3 function signatures
-const MULTICALL_SIGNATURES = {
-  // aggregate3: (Call3[] calldata calls) -> (Result[] memory returnData)
-  // where Call3 = { target, allowFailure, callData }
-  aggregate3: 'aggregate3((address,bool,bytes)[])',
-  // aggregate3Value: (Call3Value[] calldata calls) -> (Result[] memory returnData)
-  // where Call3Value = { target, allowFailure, value, callData }
-  aggregate3Value: 'aggregate3Value((address,bool,uint256,bytes)[])',
-  // tryAggregate: (bool requireSuccess, Call[] calldata calls) -> (Result[] memory returnData)
-  tryAggregate: 'tryAggregate(bool,(address,bytes)[])',
-  // aggregate: (Call[] calldata calls) -> (uint256 blockNumber, bytes[] memory returnData)
-  aggregate: 'aggregate((address,bytes)[])',
-};
-
-/**
- * Get the mailbox address for a chain from the registry
- */
-function getMailboxAddress(chainName: string): Address | undefined {
-  const addresses = chainAddresses[chainName as keyof typeof chainAddresses];
-  return (addresses as Record<string, string> | undefined)?.mailbox;
-}
-
-/**
- * Get the batch contract address for a chain from the registry (if available)
- */
-function getBatchContractAddress(chainName: string): Address | undefined {
-  const addresses = chainAddresses[chainName as keyof typeof chainAddresses];
-  return (addresses as Record<string, string> | undefined)?.batchContractAddress;
-}
-
-/**
- * Check if an address is a known multicall/batch contract
- */
-function isMulticallAddress(address: Address, chainName: string): boolean {
-  const normalizedAddress = address.toLowerCase();
-
-  // Check canonical Multicall3 address
-  if (normalizedAddress === MULTICALL3_ADDRESS.toLowerCase()) {
-    return true;
-  }
-
-  // Check chain-specific batch contract address from registry
-  const batchContract = getBatchContractAddress(chainName);
-  if (batchContract && normalizedAddress === batchContract.toLowerCase()) {
-    return true;
-  }
-
-  return false;
-}
-
-function tryDecodeMulticallCalls(txData: string): IcaCall[] | null {
-  try {
-    const selector = txData.slice(0, 10);
-
-    if (selector === utils.id(MULTICALL_SIGNATURES.aggregate3).slice(0, 10)) {
-      const decoded = utils.defaultAbiCoder.decode(
-        ['tuple(address target, bool allowFailure, bytes callData)[]'],
-        '0x' + txData.slice(10),
-      );
-      const calls = decoded[0] as Array<{
-        target: string;
-        callData: string;
-      }>;
-      return calls.map((call) => ({ to: call.target, value: '0', data: call.callData }));
-    }
-
-    if (selector === utils.id(MULTICALL_SIGNATURES.aggregate3Value).slice(0, 10)) {
-      const decoded = utils.defaultAbiCoder.decode(
-        ['tuple(address target, bool allowFailure, uint256 value, bytes callData)[]'],
-        '0x' + txData.slice(10),
-      );
-      const calls = decoded[0] as Array<{
-        target: string;
-        value: BigNumber;
-        callData: string;
-      }>;
-      return calls.map((call) => ({
-        to: call.target,
-        value: call.value.toString(),
-        data: call.callData,
-      }));
-    }
-
-    if (selector === utils.id(MULTICALL_SIGNATURES.tryAggregate).slice(0, 10)) {
-      const decoded = utils.defaultAbiCoder.decode(
-        ['bool', 'tuple(address target, bytes callData)[]'],
-        '0x' + txData.slice(10),
-      );
-      const calls = decoded[1] as Array<{ target: string; callData: string }>;
-      return calls.map((call) => ({ to: call.target, value: '0', data: call.callData }));
-    }
-
-    if (selector === utils.id(MULTICALL_SIGNATURES.aggregate).slice(0, 10)) {
-      const decoded = utils.defaultAbiCoder.decode(
-        ['tuple(address target, bytes callData)[]'],
-        '0x' + txData.slice(10),
-      );
-      const calls = decoded[0] as Array<{ target: string; callData: string }>;
-      return calls.map((call) => ({ to: call.target, value: '0', data: call.callData }));
-    }
-  } catch (error) {
-    logger.debug('Failed to decode multicall calls', error);
-  }
-
-  return null;
-}
-
-export function decodeMulticallIcaCalls(
-  call: IcaCall,
-  destinationChainName: string | undefined,
-): IcaCall[] | null {
-  if (!destinationChainName || !isMulticallAddress(call.to, destinationChainName)) return null;
-  return tryDecodeMulticallCalls(call.data);
-}
-
-const KNOWN_CALL_SELECTORS: Record<string, { name: string; sig: string }> = {
-  '0x095ea7b3': { name: 'approve', sig: 'approve(address,uint256)' },
-  '0xa9059cbb': { name: 'transfer', sig: 'transfer(address,uint256)' },
-  '0x23b872dd': { name: 'transferFrom', sig: 'transferFrom(address,address,uint256)' },
-  '0x81b4e8b4': { name: 'transferRemote', sig: 'transferRemote(uint32,bytes32,uint256)' },
-  '0x51debffc': {
-    name: 'transferRemote',
-    sig: 'transferRemote(uint32,bytes32,uint256,bytes,address)',
-  },
-  '0x24856bc3': { name: 'execute', sig: 'execute(bytes,bytes[])' },
-  '0x3593564c': { name: 'execute', sig: 'execute(bytes,bytes[],uint256)' },
-};
-
-function formatTokenAmount(amount: BigNumber): string {
-  const formatted18 = fromWei(amount.toString(), 18);
-  const parsed18 = Number.parseFloat(formatted18);
-
-  if (parsed18 > 0 && parsed18 < 0.000001) return `${amount.toString()} raw units`;
-  if (parsed18 === 0) return '0';
-  return `${formatAmountWithCommas(formatted18)} tokens`;
-}
-
-// Hyperlane universal-router Commands.sol uses a 6-bit command mask and
-// EXECUTE_SUB_PLAN = 0x21, which differs from upstream Uniswap Universal Router:
-// https://github.com/hyperlane-xyz/universal-router
-const UNIVERSAL_ROUTER_COMMAND_TYPE_MASK = 0x3f;
-const UNIVERSAL_ROUTER_V3_SWAP_EXACT_IN = 0x00;
-const UNIVERSAL_ROUTER_V3_SWAP_EXACT_OUT = 0x01;
-const UNIVERSAL_ROUTER_V2_SWAP_EXACT_IN = 0x08;
-const UNIVERSAL_ROUTER_V2_SWAP_EXACT_OUT = 0x09;
-const UNIVERSAL_ROUTER_SWEEP = 0x04;
-const UNIVERSAL_ROUTER_TRANSFER = 0x05;
-const UNIVERSAL_ROUTER_PAY_PORTION = 0x06;
-const UNIVERSAL_ROUTER_UNWRAP_WETH = 0x0c;
-const UNIVERSAL_ROUTER_EXECUTE_SUB_PLAN = 0x21;
-const NATIVE_TOKEN_SENTINEL = 'native' as const;
-
-function decodeUniversalRouterPackedPath(input: string): string | null {
-  for (const types of [
-    ['address', 'uint256', 'uint256', 'bytes', 'bool', 'bool'],
-    ['address', 'uint256', 'uint256', 'bytes', 'bool'],
-  ]) {
-    try {
-      const args = utils.defaultAbiCoder.decode(types, input);
-      const path = strip0x(args[3] as string);
-      if (path.length >= 40) return path;
-    } catch {
-      // Try the next router input shape.
-    }
-  }
-  return null;
-}
-
-function decodeUniversalRouterAddressPath(input: string): string[] | null {
-  for (const types of [
-    ['address', 'uint256', 'uint256', 'address[]', 'bool', 'bool'],
-    ['address', 'uint256', 'uint256', 'address[]', 'bool'],
-  ]) {
-    try {
-      const args = utils.defaultAbiCoder.decode(types, input);
-      const path = args[3] as string[];
-      if (path.length >= 2) return path;
-    } catch {
-      // Try the next router input shape.
-    }
-  }
-  return null;
-}
-
-function decodeUniversalRouterSwapArgs(input: string, isV2: boolean) {
-  for (const pathType of isV2 ? ['address[]'] : ['bytes']) {
-    for (const hasExtraBool of [true, false]) {
-      const types = [
-        'address',
-        'uint256',
-        'uint256',
-        pathType,
-        'bool',
-        ...(hasExtraBool ? ['bool'] : []),
-      ];
-      try {
-        return utils.defaultAbiCoder.decode(types, input);
-      } catch {
-        // Try the next router input shape.
-      }
-    }
-  }
-  return null;
-}
-
-function decodeUniversalRouterSwapCommand(
-  input: string,
-  isExactIn: boolean,
-  isV2: boolean,
-): DecodedUniversalRouterSwapCommand | null {
-  const args = decodeUniversalRouterSwapArgs(input, isV2);
-  const outputAmount = args?.[isExactIn ? 2 : 1] as BigNumber | undefined;
-
-  if (isV2) {
-    const addressPath = decodeUniversalRouterAddressPath(input);
-    if (addressPath) {
-      return {
-        tokenIn: addressPath[0],
-        tokenOut: addressPath[addressPath.length - 1],
-        outputAmount: outputAmount && !outputAmount.isZero() ? outputAmount.toString() : undefined,
-        outputAmountKind:
-          outputAmount && !outputAmount.isZero() ? (isExactIn ? 'minimum' : 'exact') : undefined,
-        recipient: args?.[0] as string | undefined,
-      };
-    }
-  }
-
-  const path = decodeUniversalRouterPackedPath(input);
-  if (!path) return null;
-
-  const route = isExactIn
-    ? { tokenIn: '0x' + path.slice(0, 40), tokenOut: '0x' + path.slice(-40) }
-    : { tokenIn: '0x' + path.slice(-40), tokenOut: '0x' + path.slice(0, 40) };
-  return {
-    ...route,
-    outputAmount: outputAmount && !outputAmount.isZero() ? outputAmount.toString() : undefined,
-    outputAmountKind:
-      outputAmount && !outputAmount.isZero() ? (isExactIn ? 'minimum' : 'exact') : undefined,
-    recipient: args?.[0] as string | undefined,
-  };
-}
-
-type DecodedUniversalRouterSwap = {
-  tokenIn: string;
-  tokenOut: string;
-  tokenOutType?: 'token' | 'native';
-  outputAmount?: string;
-  outputAmountKind?: 'exact' | 'minimum';
-  wrappedNativeToken?: string;
-  outputRecipients?: string[];
-};
-
-type DecodedUniversalRouterSwapCommand = {
-  tokenIn: string;
-  tokenOut: string;
-  outputAmount?: string;
-  outputAmountKind?: 'exact' | 'minimum';
-  recipient?: string;
-};
-
-type UniversalRouterPlanState = {
-  tokenIn?: string;
-  tokenOut?: string;
-  tokenOutType?: 'native';
-  outputAmount?: string;
-  outputAmountKind?: 'exact' | 'minimum';
-  wrappedNativeToken?: string;
-  outputRecipients: string[];
-};
-
-function decodeUniversalRouterPlan(
-  commandsBytes: string,
-  inputs: string[],
-  depth = 0,
-): DecodedUniversalRouterSwap | null {
-  if (depth > 4) return null;
-
-  const commands = strip0x(commandsBytes);
-  const state = createUniversalRouterPlanState();
-
-  for (let i = 0; i < commands.length / 2; i++) {
-    try {
-      const command = Number.parseInt(commands.slice(i * 2, i * 2 + 2), 16);
-      const commandType = command & UNIVERSAL_ROUTER_COMMAND_TYPE_MASK;
-      const input = inputs[i];
-      if (!input) continue;
-
-      if (commandType === UNIVERSAL_ROUTER_EXECUTE_SUB_PLAN) {
-        applySubPlanCommand(state, input, depth);
-        continue;
-      }
-
-      const swapCommand = getUniversalRouterSwapCommand(commandType);
-      if (swapCommand) {
-        applySwapCommand(state, input, swapCommand);
-        continue;
-      }
-
-      if (isSweepLikeCommand(commandType)) {
-        applySweepLikeCommand(state, input, commandType);
-        continue;
-      }
-
-      if (commandType === UNIVERSAL_ROUTER_UNWRAP_WETH) {
-        applyUnwrapWethCommand(state, input);
-      }
-    } catch {
-      continue;
-    }
-  }
-
-  return toDecodedUniversalRouterSwap(state);
-}
-
-function createUniversalRouterPlanState(): UniversalRouterPlanState {
-  return { outputRecipients: [] };
-}
-
-function getUniversalRouterSwapCommand(commandType: number) {
-  if (
-    commandType === UNIVERSAL_ROUTER_V3_SWAP_EXACT_IN ||
-    commandType === UNIVERSAL_ROUTER_V2_SWAP_EXACT_IN
-  ) {
-    return {
-      isExactIn: true,
-      isV2: commandType === UNIVERSAL_ROUTER_V2_SWAP_EXACT_IN,
-    };
-  }
-
-  if (
-    commandType === UNIVERSAL_ROUTER_V3_SWAP_EXACT_OUT ||
-    commandType === UNIVERSAL_ROUTER_V2_SWAP_EXACT_OUT
-  ) {
-    return {
-      isExactIn: false,
-      isV2: commandType === UNIVERSAL_ROUTER_V2_SWAP_EXACT_OUT,
-    };
-  }
-
-  return null;
-}
-
-function applySubPlanCommand(state: UniversalRouterPlanState, input: string, depth: number) {
-  const [subCommands, subInputs] = utils.defaultAbiCoder.decode(['bytes', 'bytes[]'], input) as [
-    string,
-    string[],
-  ];
-  const subSwap = decodeUniversalRouterPlan(subCommands, subInputs, depth + 1);
-  if (!subSwap) return;
-
-  state.tokenIn ??= subSwap.tokenIn;
-  state.tokenOut = subSwap.tokenOut;
-  state.tokenOutType = subSwap.tokenOutType === 'native' ? 'native' : undefined;
-  state.outputAmount = subSwap.outputAmount;
-  state.outputAmountKind = subSwap.outputAmountKind;
-  state.wrappedNativeToken = subSwap.wrappedNativeToken;
-  state.outputRecipients = subSwap.outputRecipients ?? [];
-}
-
-function applySwapCommand(
-  state: UniversalRouterPlanState,
-  input: string,
-  command: { isExactIn: boolean; isV2: boolean },
-) {
-  const decodedSwap = decodeUniversalRouterSwapCommand(input, command.isExactIn, command.isV2);
-  if (!decodedSwap) return;
-
-  state.tokenIn ??= decodedSwap.tokenIn;
-  state.tokenOut = decodedSwap.tokenOut;
-  state.tokenOutType = undefined;
-  state.outputAmount = decodedSwap.outputAmount;
-  state.outputAmountKind = decodedSwap.outputAmountKind;
-  state.wrappedNativeToken = undefined;
-  state.outputRecipients = decodedSwap.recipient ? [decodedSwap.recipient] : [];
-}
-
-function isSweepLikeCommand(commandType: number) {
-  return (
-    commandType === UNIVERSAL_ROUTER_SWEEP ||
-    commandType === UNIVERSAL_ROUTER_TRANSFER ||
-    commandType === UNIVERSAL_ROUTER_PAY_PORTION
-  );
-}
-
-function applySweepLikeCommand(
-  state: UniversalRouterPlanState,
-  input: string,
-  commandType: number,
-) {
-  if (!state.tokenIn) return;
-
-  const args = utils.defaultAbiCoder.decode(['address', 'address', 'uint256'], input);
-  const amount = args[2] as BigNumber;
-  const outputAmount =
-    (commandType === UNIVERSAL_ROUTER_TRANSFER || commandType === UNIVERSAL_ROUTER_SWEEP) &&
-    !amount.isZero()
-      ? amount.toString()
-      : undefined;
-
-  state.tokenOut = args[0] as string;
-  state.tokenOutType = undefined;
-  state.wrappedNativeToken = undefined;
-  state.outputAmount = outputAmount;
-  state.outputAmountKind =
-    commandType === UNIVERSAL_ROUTER_TRANSFER && outputAmount
-      ? 'exact'
-      : commandType === UNIVERSAL_ROUTER_SWEEP && outputAmount
-        ? 'minimum'
-        : undefined;
-  state.outputRecipients = [args[1] as string];
-}
-
-function applyUnwrapWethCommand(state: UniversalRouterPlanState, input: string) {
-  if (!state.tokenOut) return;
-
-  const args = utils.defaultAbiCoder.decode(['address', 'uint256'], input);
-  const outputAmount = !(args[1] as BigNumber).isZero()
-    ? (args[1] as BigNumber).toString()
-    : undefined;
-
-  state.wrappedNativeToken = state.tokenOut;
-  state.tokenOut = NATIVE_TOKEN_SENTINEL;
-  state.tokenOutType = 'native';
-  state.outputAmount = outputAmount;
-  state.outputAmountKind = outputAmount ? 'minimum' : undefined;
-  state.outputRecipients = [args[0] as string];
-}
-
-function toDecodedUniversalRouterSwap(
-  state: UniversalRouterPlanState,
-): DecodedUniversalRouterSwap | null {
-  if (!state.tokenIn || !state.tokenOut) return null;
-
-  return {
-    tokenIn: state.tokenIn,
-    tokenOut: state.tokenOut,
-    ...(state.tokenOutType ? { tokenOutType: state.tokenOutType } : {}),
-    ...(state.outputAmount ? { outputAmount: state.outputAmount } : {}),
-    ...(state.outputAmountKind ? { outputAmountKind: state.outputAmountKind } : {}),
-    ...(state.wrappedNativeToken ? { wrappedNativeToken: state.wrappedNativeToken } : {}),
-    ...(state.outputRecipients.length ? { outputRecipients: state.outputRecipients } : {}),
-  };
-}
-
-function decodeUniversalRouterSwap(
-  data: string,
-  functionName: 'execute(bytes,bytes[])' | 'execute(bytes,bytes[],uint256)',
-): DecodedUniversalRouterSwap | null {
-  try {
-    const routerInterface = new utils.Interface([`function ${functionName}`]);
-    const decoded = routerInterface.decodeFunctionData('execute', data);
-    return decodeUniversalRouterPlan(decoded[0] as string, decoded[1] as string[]);
-  } catch {
-    return null;
-  }
-}
-
-export function decodeIcaCallData(
-  data: string,
-  tryGetChainName?: (domainId: number) => string | undefined,
-): DecodedIcaCallData | null {
-  if (!data || data.length < 10) return null;
-
-  const selector = data.slice(0, 10).toLowerCase();
-  const known = KNOWN_CALL_SELECTORS[selector];
-  if (!known) return null;
-
-  try {
-    const iface = new utils.Interface([`function ${known.sig}`]);
-    const decoded = iface.decodeFunctionData(known.name, data);
-
-    switch (known.sig) {
-      case 'approve(address,uint256)': {
-        const spender = shortenAddress(decoded[0] as string);
-        const amount = formatTokenAmount(decoded[1] as BigNumber);
-        return {
-          functionName: 'approve',
-          summary: `Approve ${spender} to spend ${amount}`,
-        };
-      }
-      case 'transfer(address,uint256)': {
-        const recipient = shortenAddress(decoded[0] as string);
-        const amount = formatTokenAmount(decoded[1] as BigNumber);
-        return {
-          functionName: 'transfer',
-          summary: `Transfer ${amount} to ${recipient}`,
-        };
-      }
-      case 'transferFrom(address,address,uint256)': {
-        const sender = shortenAddress(decoded[0] as string);
-        const recipient = shortenAddress(decoded[1] as string);
-        const amount = formatTokenAmount(decoded[2] as BigNumber);
-        return {
-          functionName: 'transferFrom',
-          summary: `Transfer ${amount} from ${sender} to ${recipient}`,
-        };
-      }
-      case 'transferRemote(uint32,bytes32,uint256)':
-      case 'transferRemote(uint32,bytes32,uint256,bytes,address)': {
-        const destinationDomain = BigNumber.from(decoded[0]).toNumber();
-        const recipient = shortenAddress(bytes32ToAddress(decoded[1] as string));
-        const amount = formatTokenAmount(decoded[2] as BigNumber);
-        const destination = tryGetChainName?.(destinationDomain) || `domain ${destinationDomain}`;
-        return {
-          functionName: 'transferRemote',
-          summary: `Bridge ${amount} to ${recipient} on ${destination}`,
-        };
-      }
-      case 'execute(bytes,bytes[])':
-      case 'execute(bytes,bytes[],uint256)': {
-        const swap = decodeUniversalRouterSwap(data, known.sig);
-        if (!swap) return { functionName: 'execute', summary: 'Universal Router execute' };
-        const outputDisplay =
-          swap.tokenOutType === 'native' ? 'native token' : shortenAddress(swap.tokenOut);
-        return {
-          functionName: 'swap',
-          summary: `Swap ${shortenAddress(swap.tokenIn)} -> ${outputDisplay}`,
-          swap,
-          details: [
-            { label: 'Input token (origin):', value: swap.tokenIn },
-            {
-              label: 'Output token (destination):',
-              value: swap.tokenOutType === 'native' ? 'Native token' : swap.tokenOut,
-            },
-          ],
-        };
-      }
-      default:
-        return null;
-    }
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Try to extract process calls from a multicall transaction.
- * Supports various Multicall contract formats (Multicall3, etc.)
- */
-function tryDecodeMulticall(
-  txData: string,
-  mailboxInterface: utils.Interface,
-  mailboxAddress: string | undefined,
-): Array<{ metadata: string; message: string }> {
-  const results: Array<{ metadata: string; message: string }> = [];
-  if (!mailboxAddress) return results;
-  const mailboxLower = mailboxAddress.toLowerCase();
-
-  const tryParseProcessCall = (target: string, callData: string) => {
-    if (target.toLowerCase() !== mailboxLower) return;
-    try {
-      const parsed = mailboxInterface.parseTransaction({ data: callData });
-      if (parsed.name === 'process') {
-        results.push({
-          metadata: parsed.args[0] as string,
-          message: parsed.args[1] as string,
-        });
-      }
-    } catch {
-      // Not a process call, continue
-    }
-  };
-
-  try {
-    const calls = tryDecodeMulticallCalls(txData);
-    if (!calls) return results;
-    for (const call of calls) tryParseProcessCall(call.to, call.data);
-    return results;
-  } catch (error) {
-    logger.debug('Failed to decode multicall', error);
-  }
-
-  return results;
-}
-
-/**
- * Fetch the calls for a REVEAL message by parsing the process transaction.
- * The calls are passed as ISM metadata to the mailbox.process() function.
- * Handles both direct process calls to the mailbox and multicall batches.
- */
-export async function fetchRevealCalls(
-  destinationChainName: string,
-  processTxHash: string,
-  messageId: string,
-  multiProvider: ExplorerMultiProvider,
-): Promise<IcaCall[] | null> {
-  try {
-    const provider = multiProvider.getEthersV5Provider(destinationChainName);
-    const tx = await provider.getTransaction(processTxHash);
-
-    if (!tx || !tx.data || !tx.to) {
-      logger.debug('Transaction not found or has no data/to address');
-      return null;
-    }
-
-    // eslint-disable-next-line camelcase
-    const { Mailbox__factory } = await import('@hyperlane-xyz/core');
-    // eslint-disable-next-line camelcase
-    const mailboxInterface = Mailbox__factory.createInterface();
-
-    const mailboxAddress = getMailboxAddress(destinationChainName);
-    const txTo = tx.to.toLowerCase();
-
-    // Check if this is a direct call to the mailbox
-    if (mailboxAddress && txTo === mailboxAddress.toLowerCase()) {
-      logger.debug('Direct process call to mailbox detected');
-      try {
-        const decoded = mailboxInterface.parseTransaction({ data: tx.data });
-
-        if (decoded.name === 'process') {
-          const metadata = decoded.args[0] as string;
-          const revealData = decodeRevealMetadata(metadata);
-
-          if (revealData) {
-            return revealData.calls;
-          }
-        }
-      } catch {
-        logger.debug('Failed to decode direct process call');
-      }
-      return null;
-    }
-
-    // Check if this is a multicall transaction
-    if (isMulticallAddress(tx.to, destinationChainName)) {
-      logger.debug('Multicall transaction detected');
-      const processCalls = tryDecodeMulticall(tx.data, mailboxInterface, mailboxAddress);
-
-      if (processCalls.length > 0) {
-        // Find the process call that matches our message ID
-        const { messageId: computeMessageId } = await import('@hyperlane-xyz/utils');
-
-        for (const processCall of processCalls) {
-          try {
-            const msgId = computeMessageId(processCall.message);
-            if (msgId.toLowerCase() === messageId.toLowerCase()) {
-              const revealData = decodeRevealMetadata(processCall.metadata);
-              if (revealData) {
-                return revealData.calls;
-              }
-            }
-          } catch {
-            // Failed to compute message ID, continue
-          }
-        }
-
-        // If we couldn't match by message ID, return null to avoid showing potentially incorrect data
-        // (the first process call might belong to a different message in a batched transaction)
-        logger.debug('Could not match message ID, calls unavailable');
-      }
-      return null;
-    }
-
-    // Unknown destination contract - try both approaches as fallback
-    logger.debug('Unknown destination contract, trying fallback decoding');
-
-    // Try direct process call first
-    try {
-      const decoded = mailboxInterface.parseTransaction({ data: tx.data });
-      if (decoded.name === 'process') {
-        const metadata = decoded.args[0] as string;
-        const revealData = decodeRevealMetadata(metadata);
-        if (revealData) {
-          return revealData.calls;
-        }
-      }
-    } catch {
-      // Not a direct process call
-    }
-
-    // Try multicall decode
-    const processCalls = tryDecodeMulticall(tx.data, mailboxInterface, mailboxAddress);
-    if (processCalls.length > 0) {
-      const { messageId: computeMessageId } = await import('@hyperlane-xyz/utils');
-
-      for (const processCall of processCalls) {
-        try {
-          const msgId = computeMessageId(processCall.message);
-          if (msgId.toLowerCase() === messageId.toLowerCase()) {
-            const revealData = decodeRevealMetadata(processCall.metadata);
-            if (revealData) {
-              return revealData.calls;
-            }
-          }
-        } catch {
-          // Failed to compute message ID, continue
-        }
-      }
-    }
-
-    return null;
-  } catch (error) {
-    logger.error('Error fetching reveal calls', error);
-    return null;
-  }
-}
-
-/**
- * Hook to fetch the derived ICA address for a given owner on the destination chain.
- * Uses the InterchainAccountRouter's getLocalInterchainAccount function.
- *
- * The ICA address is derived from: origin domain, owner, router, ISM, and salt.
- */
-export function useIcaAddress(
-  originChainName: string | undefined,
-  destinationChainName: string | undefined,
-  owner: Address | undefined,
-  ism: Address | undefined,
-  salt: string | undefined,
-) {
-  const multiProvider = useReadyMultiProvider();
-  const chainMetadataResolver = useChainMetadataResolver();
-
-  return useQuery({
-    queryKey: [
-      'icaAddress',
-      originChainName,
-      destinationChainName,
-      owner,
-      ism,
-      salt,
-      !!multiProvider,
-    ],
-    queryFn: async () => {
-      if (
-        !originChainName ||
-        !destinationChainName ||
-        !owner ||
-        !multiProvider ||
-        !isEVMLike(
-          chainMetadataResolver.tryGetProtocol(destinationChainName) ?? ProtocolType.Unknown,
-        )
-      ) {
-        return null;
-      }
-
-      // Get the ICA router addresses for both chains
-      const originRouter = getIcaRouterAddress(originChainName);
-      const destRouter = getIcaRouterAddress(destinationChainName);
-
-      if (!originRouter || !destRouter) {
-        logger.debug('ICA router not found for chains', originChainName, destinationChainName);
-        return null;
-      }
-
-      const provider = multiProvider.getEthersV5Provider(destinationChainName);
-      const originDomainId = multiProvider.getDomainId(originChainName);
-
-      return computeIcaAddress(
-        originDomainId,
-        owner,
-        originRouter,
-        destRouter,
-        ism,
-        salt,
-        provider,
-      );
-    },
-    retry: false,
-    staleTime: 1000 * 60 * 5, // 5 minutes
-  });
-}
-
-/**
- * Hook to fetch the calls from a REVEAL message by parsing the destination transaction.
- */
-export function useRevealCalls(
-  destinationChainName: string | undefined,
-  processTxHash: string | undefined,
-  messageId: string | undefined,
-  messageType: IcaMessageType | undefined,
-) {
-  const multiProvider = useReadyMultiProvider();
-  const chainMetadataResolver = useChainMetadataResolver();
-
-  return useQuery({
-    queryKey: [
-      'revealCalls',
-      destinationChainName,
-      processTxHash,
-      messageId,
-      messageType,
-      !!multiProvider,
-    ],
-    queryFn: async () => {
-      if (
-        !destinationChainName ||
-        !processTxHash ||
-        !messageId ||
-        messageType !== IcaMessageType.REVEAL ||
-        !multiProvider ||
-        !isEVMLike(
-          chainMetadataResolver.tryGetProtocol(destinationChainName) ?? ProtocolType.Unknown,
-        )
-      ) {
-        return null;
-      }
-
-      return fetchRevealCalls(destinationChainName, processTxHash, messageId, multiProvider);
-    },
-    retry: false,
-    enabled: messageType === IcaMessageType.REVEAL && !!processTxHash && !!messageId,
-    staleTime: 1000 * 60 * 5, // 5 minutes
-  });
-}
-
-/**
- * Hook to fetch the CCIP Read ISM address and URLs for REVEAL messages.
- *
- * For REVEAL messages, the CCIP Read ISM is used to verify the commitment.
- * If the ISM in the message is zero address, we fetch the default ISM from the ICA Router.
- * The ISM's urls() function returns the off-chain gateway URLs used to fetch reveal metadata.
- */
-export function useCcipReadIsmUrls(
-  destinationChainName: string | undefined,
-  messageBytes: string | undefined,
-  messageType: IcaMessageType | undefined,
-) {
-  const multiProvider = useReadyMultiProvider();
-  const chainMetadataResolver = useChainMetadataResolver();
-
-  return useQuery({
-    queryKey: ['ccipReadIsmUrls', destinationChainName, messageBytes, messageType, !!multiProvider],
-    queryFn: async () => {
-      if (
-        !destinationChainName ||
-        !messageBytes ||
-        messageType !== IcaMessageType.REVEAL ||
-        !multiProvider ||
-        !isEVMLike(
-          chainMetadataResolver.tryGetProtocol(destinationChainName) ?? ProtocolType.Unknown,
-        )
-      ) {
-        return null;
-      }
-
-      try {
-        const provider = multiProvider.getEthersV5Provider(destinationChainName);
-
-        // Get the ICA Router address for the destination chain
-        const icaRouterAddress = getIcaRouterAddress(destinationChainName);
-        if (!icaRouterAddress) {
-          logger.debug('ICA router not found for chain', destinationChainName);
-          return null;
-        }
-
-        // Call route(message) on the ICA Router to get the ISM address for this message
-        const routerInterface = new utils.Interface([
-          'function route(bytes calldata _message) view returns (address)',
-        ]);
-        const routerContract = new Contract(icaRouterAddress, routerInterface, provider);
-        const ismAddress = await routerContract.route(messageBytes);
-
-        if (!ismAddress || ismAddress === '0x0000000000000000000000000000000000000000') {
-          return null;
-        }
-
-        // Fetch URLs from the CCIP Read ISM
-        const ismInterface = new utils.Interface(['function urls() view returns (string[])']);
-        const ismContract = new Contract(ismAddress, ismInterface, provider);
-
-        const urls = await ismContract.urls();
-        return { ismAddress, urls: urls as string[] };
-      } catch (error) {
-        logger.debug('Error fetching CCIP Read ISM URLs', error);
-        return null;
-      }
-    },
-    retry: false,
-    enabled: messageType === IcaMessageType.REVEAL && !!destinationChainName && !!messageBytes,
-    staleTime: 1000 * 60 * 5, // 5 minutes
-  });
-}
-
-/**
- * Hook to find the related ICA message (COMMITMENT <-> REVEAL) by searching
- * for messages in the same origin transaction with matching commitment hash.
- *
- * For commit-reveal flow, both COMMITMENT and REVEAL messages are dispatched
- * in the same transaction via callRemoteCommitReveal().
- */
-export function useRelatedIcaMessage(
-  originTxHash: string | undefined,
-  originDomainId: DomainId | undefined,
-  currentMsgId: string | undefined,
-  currentCommitment: string | undefined,
-  currentMessageType: IcaMessageType | undefined,
-) {
-  const { scrapedDomains } = useScrapedDomains();
-  const chainMetadataResolver = useChainMetadataResolver();
-  const multiProvider = useReadyMultiProvider();
-  const registry = useRegistry();
-  const isPiOrigin =
-    originDomainId !== undefined &&
-    isPiChain(chainMetadataResolver, scrapedDomains, originDomainId);
-
-  // Only search for related messages if this is a COMMITMENT or REVEAL message
-  const shouldSearch =
-    !!originTxHash &&
-    !!currentMsgId &&
-    !!currentCommitment &&
-    (currentMessageType === IcaMessageType.COMMITMENT ||
-      currentMessageType === IcaMessageType.REVEAL);
-
-  // Build query to fetch all messages from the same origin tx
-  // Note: We must always return a valid GraphQL query string (even when paused)
-  // because urql may validate the query before checking the pause flag
-  const { query, variables } = useMemo(() => {
-    if (!shouldSearch || !originTxHash) {
-      // Return a valid no-op query that will be paused anyway
-      return buildMessageQuery(
-        MessageIdentifierType.OriginTxHash,
-        '0x0000000000000000000000000000000000000000000000000000000000000000',
-        1,
-        true,
-      );
-    }
-    return buildMessageQuery(MessageIdentifierType.OriginTxHash, originTxHash, 100, true);
-  }, [shouldSearch, originTxHash]);
-
-  // Execute the query
-  const [{ data, fetching, error }] = useUrqlQuery<MessagesStubQueryResult>({
-    query,
-    variables,
-    pause: !shouldSearch || isPiOrigin,
-  });
-
-  const {
-    data: piMessages,
-    isFetching: isPiFetching,
-    isError: isPiError,
-  } = useQuery({
-    queryKey: ['relatedIcaPiMessages', originTxHash, originDomainId, currentMsgId],
-    queryFn: async () => {
-      if (!originTxHash || originDomainId === undefined || !multiProvider) return [];
-      const originMetadata = chainMetadataResolver.tryGetChainMetadata(originDomainId);
-      if (!originMetadata) return [];
-      return fetchMessagesFromPiChain(
-        originMetadata,
-        { input: originTxHash },
-        multiProvider,
-        registry,
-        PiQueryType.TxHash,
-      );
-    },
-    enabled: shouldSearch && isPiOrigin && !!multiProvider,
-    retry: false,
-  });
-
-  // Parse and find the related message
-  const relatedMessage = useMemo(() => {
-    if (!currentCommitment || !currentMsgId) return null;
-
-    const messages = isPiOrigin
-      ? piMessages || []
-      : parseMessageStubResult(chainMetadataResolver, scrapedDomains, data);
-
-    // Find the related message by matching commitment hash
-    for (const msg of messages) {
-      // Skip the current message
-      if (msg.msgId === currentMsgId) continue;
-
-      // Decode the message to check if it's the related COMMITMENT/REVEAL
-      const decoded = decodeIcaBody(msg.body);
-      if (!decoded || !decoded.commitment) continue;
-
-      // Check if the commitment matches
-      if (decoded.commitment.toLowerCase() === currentCommitment.toLowerCase()) {
-        // Verify it's the opposite type (COMMITMENT <-> REVEAL)
-        if (
-          (currentMessageType === IcaMessageType.COMMITMENT &&
-            decoded.messageType === IcaMessageType.REVEAL) ||
-          (currentMessageType === IcaMessageType.REVEAL &&
-            decoded.messageType === IcaMessageType.COMMITMENT)
-        ) {
-          return {
-            message: msg,
-            messageType: decoded.messageType,
-            decoded,
-          };
-        }
-      }
-    }
-
-    return null;
-  }, [
-    data,
-    piMessages,
-    isPiOrigin,
-    chainMetadataResolver,
-    scrapedDomains,
-    currentMsgId,
-    currentCommitment,
-    currentMessageType,
-  ]);
-
-  return {
-    relatedMessage: relatedMessage?.message,
-    relatedMessageType: relatedMessage?.messageType,
-    relatedDecoded: relatedMessage?.decoded,
-    isFetching: isPiOrigin ? isPiFetching : fetching,
-    isError: isPiOrigin ? isPiError : !!error,
-  };
-}
